+++ app/src/main/java/com/mofaia/mypay/app/data/repository/wallet/WalletRepository.kt
@@ -1,46 +1,48 @@
 package com.mofaia.mypay.app.data.repository.wallet
 
+import java.math.BigDecimal
+
 class WalletRepository(private val dataSource: WalletDataSource): WalletDataSource {
-    override fun creditBRL(value: Double, userId: String) {
+    override fun creditBRL(value: BigDecimal, userId: String) {
         dataSource.creditBRL(value, userId)
     }
 
-    override fun creditBRL(value: Double) {
+    override fun creditBRL(value: BigDecimal) {
         dataSource.creditBRL(value)
     }
 
-    override fun creditBitcoin(value: Double) {
+    override fun creditBitcoin(value: BigDecimal) {
         dataSource.creditBitcoin(value)
     }
 
-    override fun creditBrita(value: Double) {
+    override fun creditBrita(value: BigDecimal) {
         dataSource.creditBrita(value)
     }
 
-    override fun debitBRL(value: Double) {
+    override fun debitBRL(value: BigDecimal) {
         dataSource.debitBRL(value)
     }
 
-    override fun debitBitcoin(value: Double) {
+    override fun debitBitcoin(value: BigDecimal) {
         dataSource.debitBitcoin(value)
     }
 
-    override fun debitBrita(value: Double) {
+    override fun debitBrita(value: BigDecimal) {
         dataSource.debitBrita(value)
     }
 
     override fun getExtract() {
     }
 
-    override fun getBRLBalance(onSuccess: (Double) -> Unit, onError: () -> Unit) {
+    override fun getBRLBalance(onSuccess: (BigDecimal) -> Unit, onError: () -> Unit) {
         dataSource.getBRLBalance(onSuccess, onError)
     }
 
-    override fun getBritaBalance(onSuccess: (Double) -> Unit, onError: () -> Unit) {
+    override fun getBritaBalance(onSuccess: (BigDecimal) -> Unit, onError: () -> Unit) {
         dataSource.getBritaBalance(onSuccess, onError)
     }
 
-    override fun getBiticoinBalance(onSuccess: (Double) -> Unit, onError: () -> Unit) {
+    override fun getBiticoinBalance(onSuccess: (BigDecimal) -> Unit, onError: () -> Unit) {
         dataSource.getBiticoinBalance(onSuccess, onError)
     }
